requirements.txt
playwright
pytest
openai
python-dotenv
